requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautiful-date==2.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
dateparser==1.0.0
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.3
django-multiselectfield==0.1.12
django-nested-inline==0.4.4
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
google-api-core==2.0.0
google-api-python-client==2.17.0
google-apis-oauth-django==0.1.0
google-auth==2.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
humanize==3.11.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
maya==0.6.1
mercadopago==2.0.7
numpy==1.21.2
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
pandas==1.3.2
pendulum==2.1.2
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
python-http-client==3.3.3
pytz==2021.1
pytzdata==2020.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sendgrid==6.8.2
simplejson==3.17.3
six==1.16.0
snaptime==0.2.4
sqlparse==0.4.2
starkbank-ecdsa==2.0.0
toml==0.10.2
tzlocal==2.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
